Fix permission on sites/default and settings.php during cleanup
Drupal removes write permission on the sites/default directory and settings.php
file.  When drush-deploy goes to clean it up it fails leaving dirty releases
hanging out.  The simple approach is to set the permission back prior to clean
up.  I choose to blanket set the user permsions to write (u+w) for the entire
release directory.  I felt it wasn't necessary to be specifically cautious
because the next task is to simply rm -rf the directory.  The point is to ditch
it.

This patch is an effort to clean up issue #14.